% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxpDirichlet.R
\name{approxpDirichlet}
\alias{approxpDirichlet}
\title{Approximate CDF of Dirichlet}
\usage{
approxpDirichlet(a, t, N = 10000)
}
\arguments{
\item{a}{Dirichlet parameters}

\item{t}{the proportions}

\item{N}{number of samples to draw}
}
\description{
A monte-carlo approximation of the Dirichlet CDF.
}
\examples{
approxpDirichlet(c(1,3,1), c(0.299, 0.528, 0.204))

}
\references{
\url{http://stats.stackexchange.com/questions/57262/implementation-of-dirichlet-cdf}
}
\author{
\href{http://stats.stackexchange.com/users/9394/zen}{Zen}, Neal Fultz
}
