% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag.R
\name{bag}
\alias{bag}
\title{Multi-indicators / "Bag o Words"}
\usage{
bag(..., prefix = ".", levels = NULL, `NA` = NULL)
}
\arguments{
\item{...}{the columns to bag}

\item{prefix}{a prefix for the column names}

\item{levels}{levels shared among all columns}
}
\value{
a n*p indicator matrix
}
\description{
This creates an indicator matrix from several columns.
}
\examples{

}
\references{
\url{https://stackoverflow.com/questions/47055856/search-multiple-columns-for-string-to-set-indicator-variable/57381877#57381877}
}
\author{
\href{https://stackoverflow.com/users/986793/neal-fultz}{Neal Fultz}
}
