% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_passes.R
\docType{data}
\name{bat_passes}
\alias{bat_passes}
\title{Bat passes}
\format{A data.frame with 80 observations and 5 variables.

\describe{
  \item{Location}{Five locations in the study}
  \item{Al.N}{Dark or Light condition}
  \item{Buzzes}{Count of buzzes per day}
  \item{Passes}{Count of passes per day}
  \item{Date}{Date of observation}
}}
\source{
\url{https://stats.stackexchange.com/q/325334}
}
\usage{
data(bat_passes)
}
\description{
Data from a study on the effect of light on bats.
}
\examples{
data(bat_passes)
head(bat_passes)
}
\references{
nausicaa (\url{https://stats.stackexchange.com/users/190274/nausicaa}), 
poisson glm to observe whether effects of artificial light on the number of bat passes in each location were significant, 
URL (version: 2018-03-09): \url{https://stats.stackexchange.com/q/325334}
}
\keyword{datasets}
