% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsearch7.R
\name{bsearch7}
\alias{bsearch7}
\title{Efficient binary search for character vectors}
\usage{
bsearch7(val, tab, L = 1L, H = length(tab))
}
\arguments{
\item{val}{values}

\item{tab}{table to find values in}

\item{L}{lower bound}

\item{H}{upper bound}
}
\description{
Efficient binary search for character vectors
}
\examples{
 bsearch7(sample(letters, 5000, replace=TRUE), letters)
}
\references{
\url{http://stackoverflow.com/questions/20133344/find-closest-value-in-a-vector-with-binary-search/} and
            \url{https://stat.ethz.ch/pipermail/r-help/2011-April/274182.html}
}
\author{
Martin Morgan, Neal Fultz
}
