% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBMI.R
\name{calcBMI}
\alias{calcBMI}
\title{Calculate Body Mass Index}
\usage{
calcBMI(w = 204, f = 6, i = 1)
}
\arguments{
\item{w}{Weight (in pounds)}

\item{f}{Height (feet)}

\item{i}{Height (inches)}
}
\value{
BMI
}
\description{
This calculates Body Mass Index
}
\examples{
calcBMI(199, 5, 9)
}
\references{
\url{https://stackoverflow.com/questions/16782598/declaring-dynamic-variable-in-r/16782661#16782661}
}
\seealso{
\code{\link{sprintf}}
}
\author{
\href{https://stackoverflow.com/users/190277/ben-bolker}{Ben Bolker}
}
