% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chunk2.R
\docType{data}
\name{chunk2}
\alias{chunk2}
\title{Split a vector into n chunks}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
chunk2
}
\arguments{
\item{x}{a vector}

\item{n}{number of chunks}
}
\description{
Split a vector into n chunks
}
\examples{
chunk2(1:30, 6)
}
\author{
mathheadinclouds, Dis Shishkov
}
\references{
\url{http://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r}
}
\keyword{datasets}

