% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamp.R
\name{clamp}
\alias{clamp}
\title{Clamp a value into a range}
\usage{
clamp(x, e1, e2 = -e1)
}
\arguments{
\item{x}{vector}

\item{e1}{the first edge}

\item{e2}{the other edge, defaults to the negation of e1.}
}
\value{
x, with values outside the boundaries replaced with the boundary points.
}
\description{
Splits paths into folders.
}
\examples{

clamp(-10:10, 2, -2)
clamp(-10:10, -2)
clamp(-10:10, 2)

}
\references{
\url{https://stackoverflow.com/questions/32599695/clamp-variable-within-range}
}
\author{
\href{https://stackoverflow.com/users/3093387/josliber}{josliber},
}
