% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{classMethods}
\alias{classMethods}
\title{List all methods for an object}
\usage{
classMethods(cl)
}
\arguments{
\item{cl}{a vector of class names, or an object}
}
\description{
The built-in methods() function will give all available methods for a specified class,
or for a specified generic function, but not for an object. Objects can have multiple 
classes, so this can be complicated to calculate.
}
\examples{
g <- glm(y~x,data=data.frame(x=1:10,y=1:10))
classMethods(g)

}
\references{
\url{http://stackoverflow.com/questions/23840404/function-to-return-all-s3-methods-applicable-to-an-object}
}
\author{
\href{http://stackoverflow.com/users/2372064/mrflick}{MrFlick}
}
