% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Comment.R
\docType{data}
\name{Comment}
\alias{Comment}
\title{Multi-line Comments}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
Comment
}
\arguments{
\item{...}{comment, not evaluated.}
}
\description{
Multi-line Comments
}
\examples{
Comment( `

# Put anything in here except back-ticks.

api_idea <- function() {
  return TRUE
}

# Just to show api_idea isn't really there...
print( api_idea )

`)
}
\author{
thell, Neal Fultz
}
\references{
\url{http://stackoverflow.com/questions/1231195/multiline-comment-workarounds}
}
\keyword{datasets}

