% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copyEnv.R
\name{copyEnv}
\alias{copyEnv}
\title{Copy objects from one environment to another}
\usage{
copyEnv(from, to, names = ls(from, all.names = TRUE))
}
\arguments{
\item{from}{source environment}

\item{to}{target environment}

\item{names}{names of objects to copy}
}
\description{
Copy objects from one environment to another
}
\examples{
e1 <- new.env()
e2 <- new.env()
e1$a <- 1
e1$b <- 2
copyEnv(e1,e2)
ls(e2)
}
\author{
Neal Fultz
}
\references{
\url{http://stackoverflow.com/a/11813377/986793}
}

