% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurryF.R
\name{CurryF}
\alias{CurryF}
\title{Curry a function}
\usage{
CurryF(FUN, ...)
}
\arguments{
\item{FUN}{a function}

\item{...}{arguments to apply}
}
\value{
a partially applied function
}
\description{
Partially applies a function
}
\examples{
 x <- 1:2
 appender <- CurryF(append, x)
 x <- 5:6
 append <- sum
 appender(6)
}
\references{
\url{https://stackoverflow.com/a/31900149/986793}
}
\author{
\href{https://stackoverflow.com/users/1756702/a-webb}{A Webb}
}
