% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated2.R
\name{duplicated2}
\alias{duplicated2}
\title{Find duplicates in a vector}
\usage{
duplicated2(x)
}
\arguments{
\item{x}{a vector}
}
\description{
This will find all duplicates in a run, unlike \code{\link{duplicated}} which
finds duplicates globally.
}
\examples{
duplicated2(c(2,3,3,2,2,3,3,3,3,2,2))
}
\references{
\url{http://stackoverflow.com/questions/30260507/exclude-subsequent-duplicated-rows-in-r}
}
\author{
\href{http://stackoverflow.com/users/980833/josh-obrien}{Josh O'Brien}, Neal Fultz
}
