% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.knitr.in.progress.R
\name{is.knitr.in.progress}
\alias{is.knitr.in.progress}
\title{Is knitr in progress?}
\usage{
is.knitr.in.progress()
}
\value{
TRUE if knitr is executing
}
\description{
Is knitr in progress?
}
\references{
\url{https://stackoverflow.com/questions/33107908/how-to-tell-if-code-is-executed-within-a-knitr-rmarkdown-context}
}
\author{
Yihui Xie and
  \href{https://stackoverflow.com/users/2706569/cl}{CL},
}
