% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_inst.R
\name{is_inst}
\alias{is_inst}
\title{Check if package is available}
\usage{
is_inst(pkg)
}
\arguments{
\item{pkg}{a character string with the name of a single package. An error occurs if more than one package name is given.}
}
\value{
\code{TRUE} if a package is installed, and \code{FALSE} otherwise.
}
\description{
A predicate for whether a package is installed
}
\examples{

is_inst("grDevices")

}
\references{
\url{https://stackoverflow.com/questions/9341635/check-for-installed-packages-before-running-install-packages/38082613#38082613}
}
\author{
\href{https://stackoverflow.com/users/1863950/artem-klevtsov}{Artem Klevtsov}
}
