% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lsos.R
\docType{data}
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
lsos
}
\arguments{
\item{...}{to be passed along to internal}

\item{n}{to be given to head}
}
\description{
Improved list of objects
}
\author{
Dirk Eddelbuettel
}
\references{
\url{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}
\keyword{datasets}

