% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Mode.R
\docType{data}
\name{Mode}
\alias{Mode}
\title{Calculate mode (most common element) of a vector}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
Mode
}
\arguments{
\item{x}{a vector}
}
\description{
Calculate mode (most common element) of a vector
}
\author{
Ken Williams
}
\references{
\url{http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode}
}
\keyword{datasets}

