% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseldap.R
\name{parseLDAP}
\alias{parseLDAP}
\title{Parse LDAP output into dataframe}
\usage{
parseLDAP(ldapraw)
}
\arguments{
\item{ldapraw}{A length-one character vector containing the raw LDAP output}
}
\value{
a data.frame with one row per person
}
\description{
Parse LDAP output into dataframe
}
\references{
\url{https://stackoverflow.com/questions/22793855/how-do-i-run-a-ldap-query-using-r}
}
\author{
\href{https://stackoverflow.com/users/3792484/user3792484}{user3792484},
  rewrite by Neal Fultz
}
