% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomRows.R
\name{randomRows}
\alias{randomRows}
\title{Sample rows from a dataframe or matrix}
\usage{
randomRows(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a data frame or matrix}

\item{size}{}

\item{replace}{}

\item{prob}{}
}
\description{
Sample rows from a dataframe or matrix
}
\section{Changes}{

Matched parameters to sample -- njf, May 18, 2015
}
\author{
\href{http://stackoverflow.com/users/211116/spacedman}{Spacedman}
}
\references{
\url{http://stackoverflow.com/questions/8273313/random-rows-in-dataframe-in-r}
}
\seealso{
\code{\link{sample}}
}

