% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcweibull.R
\name{rcweibull}
\alias{rcweibull}
\title{Sample from conditional Weibull}
\usage{
rcweibull(n, lambda, kappa, T0)
}
\arguments{
\item{n}{Number of samples to draw}

\item{lambda}{Weibull parameter}

\item{kappa}{Weibull parameter}

\item{T0}{Left boundary}
}
\description{
The conditional weibull distribution is a truncated version  - the condition is that the observation has already survived
}
\examples{

rcweibull(n=20, lambda=1:2, kappa=1:5, T0=1:20) 

}
\references{
\url{hhttps://stats.stackexchange.com/questions/470058/how-to-draw-from-conditional-weibull-distribution/470064#470064}
}
\seealso{
\code{\link[stats]{rweibull}}
}
\author{
\href{https://stats.stackexchange.com/users/283201/jcken}{jcken}, 
  Neal Fultz
}
