% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.directory.R
\name{read.directory}
\alias{read.directory}
\title{Bulk import data files}
\usage{
read.directory(
  path = ".",
  pattern = NULL,
  reader = read.csv,
  ...,
  reducer = function(dfs) do.call(rbind.data.frame, dfs),
  recursive = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names}

\item{pattern}{an optional \link[=regex]{regular expression}. Only file names which match the regular expression will be returned.}

\item{reader}{a function that can read data from a file name.}

\item{...}{optional arguments to pass to the reader function (eg \code{stringsAsFactors}).}

\item{reducer}{a function to unnest the individual data files. Use I to retain the nested structure.}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\description{
Replaces NA elements of x with corresponding element of y, and NA elements of 
that with corresponding element from dots.
}
\section{Changes}{


Rather than using eagerly evaluating the dot arguments and Reducing over them,
instead we use recursion to evaluate them lazily.
}

\references{
\url{https://stackoverflow.com/a/19254510/986793}
}
\author{
\href{https://stackoverflow.com/users/903061/gregor}{Gregor Thomas},
}
