% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readkey.R
\docType{data}
\name{readkey}
\alias{readkey}
\title{Wait for a keypress}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
readkey
}
\arguments{
\item{prompt}{the text to display}
}
\description{
Wait for a keypress
}
\section{Changed Feb 23, 2015 by njf}{

\code{prompt} may be set by a parameter rather than hard coding it.
}
\author{
nnn, arulmr, Neal Fultz
}
\references{
\url{http://stackoverflow.com/questions/15272916/how-to-wait-for-a-keypress-in-r}
}
\keyword{datasets}

