% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{reflect_triangle}
\alias{reflect_triangle}
\title{Reflect upper/lower triangle across diagonal}
\usage{
reflect_triangle(m, from = c("lower", "upper"))
}
\arguments{
\item{m}{a square matrix}

\item{from}{lower or upper triangle}
}
\value{
a symmetric square matrix
}
\description{
Create a new matrix by copying the lower(upper) triangle to the other half.
}
\examples{
  x <- matrix(1:9,3,3)
  reflect_triangle(x, "lower")
  reflect_triangle(x, "upper")
}
\references{
\url{https://stackoverflow.com/questions/26166569/copy-upper-triangle-to-lower-triangle-for-several-matrices-in-a-list}
}
\author{
\href{https://stackoverflow.com/users/980833/josh-obrien}{Josh O'Brien}
}
