% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null_recursively.R
\name{replace_null_recursively}
\alias{replace_null_recursively}
\title{Replace NULLs in nested lists}
\usage{
replace_null_recursively(x, what = NA_character_)
}
\arguments{
\item{x}{a nested list}

\item{what}{a value}
}
\value{
x with NULLs replaced with what
}
\description{
Replace NULLs in nested lists
}
\references{
\url{https://stackoverflow.com/a/38950427/986793}
}
\author{
\href{https://stackoverflow.com/users/6621998/shayaa}{shayaa},
}
