% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{Resave a session}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{...}{symbols of objects}

\item{list}{a character vector of object names; unfortunately named}

\item{file}{the file to update}
}
\description{
Resave a session
}
\references{
\url{http://stackoverflow.com/a/11813377/986793}
}
\seealso{
\code{\link[base]{load}}, \code{\link[base]{save}}
}
\author{
Neal Fultz and
  \href{http://stackoverflow.com/users/1201032/flodel}{flodel},
}
