% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsplit.R
\name{rsplit}
\alias{rsplit}
\title{Recursivly split a data.frame}
\usage{
rsplit(x, by, drop = FALSE)
}
\arguments{
\item{x}{a data.frame or vector}

\item{by}{a data.frame of factors}

\item{drop}{drop unused factor levels}
}
\value{
a nested list of dataframes, split by each element of \code{by}


Inspired by, but different from the below
}
\description{
When there are multiple factors to split by, Base R split returns a 
flattened structure by splitting on the interaction of all factors. 
rsplit instead returns a nested list-of-lists.
}
\references{
\url{https://stackoverflow.com/questions/47802545/converting-data-frame-into-deeply-nested-list/47802935#47802935}
}
\author{
Neal Fultz
}
