% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_path.R
\name{split_path}
\alias{split_path}
\title{Split paths into folders}
\usage{
split_path(x)
}
\arguments{
\item{x}{character vector of file paths}
}
\description{
Splits paths into folders.
}
\examples{

split_path("~")

}
\references{
\url{https://stackoverflow.com/questions/29214932/split-a-file-path-into-folder-names-vector/29232017#29232017}
}
\author{
\href{https://stackoverflow.com/users/269476/james}{James}, Neal Fultz for vectorized version
}
