% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_named.R
\name{sprintf_named}
\alias{sprintf_named}
\title{sprintf, with named references}
\usage{
sprintf_named(fmt, ...)
}
\arguments{
\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{values to be interpolated, optionally with names.}
}
\value{
a character vector.
}
\description{
This converts named references in a format string (marked by curly braces), and passes through to \code{\link{sprintf}}.
}
\examples{
sprintf_named("\%{HIA}s!!! \%{RYLAH}s", RYLAH="Rock You Like a Hurricane", HIA="Here I Am")
}
\references{
\url{https://stackoverflow.com/questions/17475803/sprintf-format-strings-reference-by-name/55423080#55423080}
}
\seealso{
\code{\link{sprintf}}
}
\author{
Neal Fultz
}
