% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackoverflow.R
\docType{package}
\name{stackoverflow}
\alias{stackoverflow}
\alias{stackoverflow-package}
\title{Stack Overflow's Greatest Hits}
\description{
The stackoverflow package consists of helper functions 
 collected from StackOverflow.com, a question and answer site for professional
 and enthusiast programmers.
}
\references{
\url{http://stackoverflow.com},
\url{https://github.com/nfultz/stackoverflow}
}
