% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strReverse.R
\docType{data}
\name{strReverse}
\alias{strReverse}
\title{Reverse each string of a vector}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
strReverse
}
\arguments{
\item{x}{a character vector}
}
\description{
A function which will reverse every string in a vector of strings.
}
\examples{
strReverse(c("abc", "Statistics"))
}
\author{
Josh O'Brien
}
\references{
\url{https://stackoverflow.com/questions/13612967/how-to-reverse-a-string-in-r}
}
\keyword{datasets}

