% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substituteExpr.R
\name{substituteExpr}
\alias{substituteExpr}
\title{Substitute on an expression in a value}
\usage{
substituteExpr(expr, env)
}
\arguments{
\item{expr}{an expression value}

\item{env}{an environment or a list object.}
}
\description{
If expr's value is an expression, substitute in any variables bound in \code{env}.
}
\details{
Differs in that substitute uses expr's expression and not value.
}
\examples{
a <- expression(z = y + x + 2)
substituteExpr(a, list(x=4))
}
\references{
\url{https://stackoverflow.com/questions/47780150/use-variable-in-r-substitute/986793}
}
\seealso{
\link{substitute}
}
\author{
\href{https://stackoverflow.com/users/516548/g-grothendieck}{G. Grothendieck}
}
