% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horner.R
\name{trim_trailing}
\alias{trim_trailing}
\title{Strip leading / trailing zeros}
\usage{
trim_trailing(x, value = 0)
}
\arguments{
\item{x}{a vector}

\item{value}{a value to strip from x}
}
\value{
a new vector, with values at the right removed
}
\description{
Removes \code{value} from rightmost elements of a vector.
}
