% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_columns.R
\name{unique_columns}
\alias{unique_columns}
\title{Remove duplicated columns}
\usage{
unique_columns(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame without duplicated columns
}
\description{
Drops duplicated columns from a data.frame (or other list-like object).
}
\examples{

df <- data.frame(a=1:10, b=1:10, c=2:11)

unique_columns(df)

}
\references{
\url{https://stackoverflow.com/a/58475153/986793}
}
\author{
\href{https://stackoverflow.com/users/3732271/akrun}{akrun}
}
