% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip2}
\alias{zip2}
\alias{enumerate}
\title{Zip / Enumerate from python}
\usage{
zip2(...)

enumerate(...)
}
\arguments{
\item{...}{Objects to be zipped together.}
}
\value{
a list of lists
}
\description{
\code{zip2}s together parallel lists into a list-of-lists. It is named zip2 to not collide with utils.
}
\details{
\code{enumerate} zips together a list with it's indices.
}
\examples{
zip2(1:5,1:10)
enumerate(l=LETTERS)
}
\references{
\url{https://stackoverflow.com/questions/9281323/zip-or-enumerate-in-r/57564884#57564884}
}
\author{
\href{https://stackoverflow.com/users/986793/neal-fultz}{Neal Fultz}
}
