% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chunk2.R
\name{chunk2}
\alias{chunk2}
\title{Split a vector into n chunks}
\usage{
chunk2(x, n)
}
\arguments{
\item{x}{a vector}

\item{n}{number of chunks}
}
\description{
Split a vector into n chunks
}
\examples{
chunk2(1:30, 6)
}
\author{
\href{http://stackoverflow.com/users/1563634/mathheadinclouds}{mathheadinclouds},
  \href{http://stackoverflow.com/users/1737569/dis-shishkov}{Dis Shishkov}
}
\references{
\url{http://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r}
}

