% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comment.R
\name{Comment}
\alias{Comment}
\title{Multi-line Comments}
\usage{
Comment(...)
}
\arguments{
\item{...}{comment, not evaluated.}
}
\description{
Multi-line Comments
}
\examples{
Comment( `

# Put anything in here except back-ticks.

api_idea <- function() {
  return TRUE
}

# Just to show api_idea isn't really there...
print( api_idea )

`)

}
\references{
\url{http://stackoverflow.com/questions/1231195/multiline-comment-workarounds}
}
\author{
\href{http://stackoverflow.com/users/173985/thell}{Thell},
\href{http://stackoverflow.com/users/211116/spacedman}{Spacedman}
  Neal Fultz
}
