% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyEnv.R
\name{copyEnv}
\alias{copyEnv}
\title{Copy objects from one environment to another}
\usage{
copyEnv(from, to, names = ls(from, all.names = TRUE))
}
\arguments{
\item{from}{source environment}

\item{to}{target environment}

\item{names}{names of objects to copy}
}
\description{
Copy objects from one environment to another
}
\examples{

e1 <- list2env(list(a=1, b=2))
e2 <- new.env()
copyEnv(e1,e2)
ls(e2)
}
\references{
\url{http://stackoverflow.com/a/33465113/986793}
}
\author{
Neal Fultz
}
