% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flatten2.R
\docType{data}
\name{flatten2}
\alias{flatten2}
\title{Flatten a list without type coercion}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 58 5 70 5 5 5 60 72
  .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x2d03fc8> 
 - attr(*, "class")= chr "autologger"
}}
\usage{
flatten2
}
\arguments{
\item{x}{a nested list}

\item{len}{guess of output length}
}
\description{
Flatten a list without type coercion
}
\section{Changed Feb 19, 2015 by njf}{

Rather than calculating length, preallocate more than needed.
}
\author{
Tommy, Joshua Ulrich, Josh O'Brien, Neal Fultz
}
\references{
\url{http://stackoverflow.com/questions/8139677/how-to-flatten-a-list-to-a-list-without-coercion}
}
\keyword{datasets}

