% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{is.rstudio}
\alias{is.rstudio}
\title{Is this in Rstudio?}
\usage{
is.rstudio()
}
\value{
TRUE if running in RStudio child process, or Rstudio console specifically.
}
\description{
Tests if inside RStudio.
}
\examples{

is.rstudio() && is.rstudio.console()

}
\references{
\url{https://stackoverflow.com/a/17804414/986793}
}
\author{
\href{https://stackoverflow.com/users/1345455/coatless}{coatless}, \href{https://stackoverflow.com/users/946850/krlmlr}{krlmr}
}
