% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_kmeans.R
\name{logLik.kmeans}
\alias{logLik.kmeans}
\title{Log-Likelihood for k-means clustering (for calculating AIC and BIC)}
\usage{
\method{logLik}{kmeans}(object, ...)
}
\arguments{
\item{object}{a \code{kmeans} object}

\item{...}{unused}
}
\description{
Log-Likelihood for k-means clustering (for calculating AIC and BIC)
}
\examples{
cl <- kmeans(iris[-5], 3)
AIC(cl)
}
\references{
\url{http://stackoverflow.com/questions/15839774/how-to-calculate-bic-for-k-means-clustering-in-r}
}
\seealso{
\code{\link[stats]{logLik}}, \code{\link[stats]{AIC}}, \code{\link[stats]{BIC}}
}
\author{
Neal Fultz, inspired by Sherry Towers and 
  \href{http://stackoverflow.com/users/2514568/andy-clifton}{Andy Clifton},
}
