% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
lsos(..., n = 10)
}
\description{
Improved list of objects
}
\author{
Dirk Eddelbuettel
}
\references{
\url{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}

