% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{to be passed along to internal}

\item{n}{to be given to head}
}
\description{
Improved list of objects
}
\author{
\href{http://stackoverflow.com/users/143305/dirk-eddelbuettel}{Dirk Eddelbuettel}
}
\references{
\url{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}

