% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Calculate mode (most common element) of a vector}
\usage{
Mode(x, ux = unique(x))
}
\arguments{
\item{x}{a vector}

\item{ux}{vector of values x may take}
}
\description{
Calculate mode (most common element) of a vector
}
\section{Changes}{

Factored \code{ux} into argument -- njf, May 18, 2015
}

\references{
\url{http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode}
}
\author{
\href{http://stackoverflow.com/users/169947/ken-williams}{Ken Williams}
}
