% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Calculate mode (most common element) of a vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Calculate mode (most common element) of a vector
}
\author{
\href{http://stackoverflow.com/users/169947/ken-williams}{Ken Williams}
}
\references{
\url{http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode}
}

