% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRows.R
\name{randomRows}
\alias{randomRows}
\title{Sample rows from a dataframe or matrix}
\usage{
randomRows(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a data frame or matrix}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements of the vector being sampled.}
}
\description{
Sample rows from a dataframe or matrix
}
\section{Changes}{

Matched parameters to sample -- njf, May 18, 2015
}

\references{
\url{http://stackoverflow.com/questions/8273313/random-rows-in-dataframe-in-r}
}
\seealso{
\code{\link{sample}}

\code{\link[dplyr]{sample_n}} for dplyr users
}
\author{
\href{http://stackoverflow.com/users/211116/spacedman}{Spacedman}
}
