% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readkey.R
\name{readkey}
\alias{readkey}
\title{Wait for a keypress}
\usage{
readkey()
}
\description{
Wait for a keypress
}
\author{
nnn, arulmr
}
\references{
\url{http://stackoverflow.com/questions/15272916/how-to-wait-for-a-keypress-in-r}
}

