% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readkey.R
\name{readkey}
\alias{readkey}
\title{Wait for a keypress}
\usage{
readkey(prompt = "Press [enter] to continue")
}
\arguments{
\item{prompt}{the text to display}
}
\description{
Wait for a keypress
}
\section{Changed Feb 23, 2015 by njf}{

\code{prompt} may be set by a parameter rather than hard coding it.
}

\references{
\url{http://stackoverflow.com/questions/15272916/how-to-wait-for-a-keypress-in-r}
}
\author{
\href{http://stackoverflow.com/users/2427707/nnn}{nnn}, 
  \href{http://stackoverflow.com/users/1095090/arulmr}{arulmr}, 
  Neal Fultz
}
