% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strReverse.R
\name{strReverse}
\alias{strReverse}
\title{Reverse each string of a vector}
\usage{
strReverse(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
A function which will reverse every string in a vector of strings.
}
\examples{
strReverse(c("abc", "Statistics"))
}
\author{
\href{http://stackoverflow.com/users/980833/josh-obrien}{Josh O'Brien}
}
\references{
\url{https://stackoverflow.com/questions/13612967/how-to-reverse-a-string-in-r}
}

