% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_trailing.R
\name{trim_trailing}
\alias{trim_trailing}
\alias{trim_leading}
\title{Strip leading / trailing zeros}
\usage{
trim_trailing(x, value = 0)

trim_leading(x, value = 0)
}
\arguments{
\item{x}{a vector}

\item{value}{a value to strip from x}
}
\value{
a new vector, with values at the right removed
}
\description{
Removes \code{value} from rightmost/leftmost elements of a vector.
}
\examples{
trim_leading(c(0,0,0,0,1:5))

}
\references{
\url{https://stackoverflow.com/questions/24009982/remove-zeros-in-the-start-and-end-of-a-vector/}
}
\author{
\href{https://stackoverflow.com/users/986793/neal-fultz}{Neal Fultz}
}
